+++ src/main/java/com/nhom7/coworkingspace/service/OtpService.java
@@ -5,4 +5,8 @@ public interface OtpService {
     void sendConfirmationOtp(String email);
 
     void sendPasswordResetOtp(String email);
+
+    void confirmAccount(String email, String otp);
+
+    void resetPassword(String email, String otp, String newPassword);
 }